+++ internal/sysutil/syslog_others.go
@@ -1,14 +1,13 @@
-// +build !windows,!nacl,!plan9
+//+build !windows,!nacl,!plan9
 
-package util
+package sysutil
 
 import (
 	"log"
 	"log/syslog"
 )
 
-// ConfigureSyslog reroutes standard logger output to syslog
-func ConfigureSyslog(serviceName string) error {
+func configureSyslog(serviceName string) error {
 	w, err := syslog.New(syslog.LOG_NOTICE|syslog.LOG_USER, serviceName)
 	if err != nil {
 		return err
